refactor: completely remove `defaul.rs`
